Add top_bar location for Zendesk Sell (#269)
This change adds `top_bar` location for Sell product. Together with location itself it also adds possibility to set icons for that location (i.e. to change the icon upon a notification).

The changes in the code are straightforward and in-line with what has been already there for Support product.

The tests for icon location have been refactored a bit to have a clearer way of mocking manifest locations differently for different scenarios.

JIRA issues: SELCHA-14, SELCHA-9